+++ recipes/head_node_base_config.rb
@@ -2,7 +2,7 @@
 
 #
 # Cookbook Name:: aws-parallelcluster
-# Recipe:: master_base_config
+# Recipe:: head_node_base_config
 #
 # Copyright 2013-2015 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 #
@@ -141,8 +141,8 @@ nfs_export "/opt/intel" do
   only_if { ::File.directory?("/opt/intel") }
 end
 
-# Setup RAID array on master node
-include_recipe 'aws-parallelcluster::setup_raid_on_master'
+# Setup RAID array on head node
+include_recipe 'aws-parallelcluster::setup_raid_on_head_node'
 
 # Configure Ganglia
 include_recipe 'aws-parallelcluster::ganglia_config'
@@ -200,6 +200,6 @@ template '/etc/sqswatcher.cfg' do
 end
 
 if node['cfncluster']['dcv_enabled'] == "master"
-  # Activate DCV on master Node
+  # Activate DCV on head node
   include_recipe 'aws-parallelcluster::dcv_config'
 end