+++ CHANGELOG.md
@@ -8,9 +8,11 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 
 **CHANGES**
 - Upgrade NICE DCV to version 2020.2-9508
+- Use inclusive language in recipe names and internal naming convention.
 
 **BUG FIXES**
 - Fix installation of Intel PSXE package on CentOS 7 by using yum4.
+- Fix routing issues with multiple Network Interfaces on Ubuntu 18.04.
 
 2.10.0
 ------