+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 2.10.1
 ------
 
+**ENHANCEMENTS**
+- Configure NFS threads to be max(8, num_cores) for performance
+
 **CHANGES**
 - Upgrade NICE DCV to version 2020.2-9508
 - Use inclusive language in recipe names and internal naming convention.